+++ conreality_headset/CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.2 - 2019-07-11
+### Added
+- `Headset.canSpeak` static property
+- `Headset.speak()` static method
+- `Headset.stopSpeaking()` static method
+
 ## 0.1.1 - 2019-07-11
 ### Added
 - Bluetooth HSP/HFP headset support (on Android only)